Update index.html
I have added my code for the remaining tasks. I also added the code for the navigation bar on each step, hopefully, it'll come out fine. 
Unfortunately, I couldn't run it to get a glimpse of what it will look like because of the multiplicity of codes, and it won't run on my end.