Update bootstrap.css
I have added some of my CSS to the file. I hope it works as well as I did get to run it.  
I also wanted to add that you may need to add the home page to the main Html file to harmonize the document (just thinking...)
Ps. I noticed I made a mistake with the name of the home file on the navigation bar; it should be "indexHome.html" and not "indexhome.html". 
Please help to make the adjustments as I can't make any changes until you agree to merge files.